Reorganize actor module
