+++ crates/common/src/actor/indicators.rs
@@ -13,6 +13,11 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
+// Under development
+#![allow(dead_code)]
+#![allow(unused_variables)]
+#![allow(unused_imports)]
+
 use std::{collections::HashMap, sync::Arc};
 
 use nautilus_indicators::indicator::Indicator;
